Fix filename too long error when passing many targets (#6)
When parsing the targets parameter, the tool tries to stat the
argument to check whether it should treat it as filename for a targets
file.
The parameter was only considered a comma separated targets list if stat
returned an IsNotExist-errors. In case of other stat errors (like
filename too long), the tool still tried to open the targets list as
file.